requirements.txt
python-whois
ipwhois
httpx
re
dnspython
requests
socket
ssl
datetime
